mq/tests/requirements.txt
pytest==7.4.3
psycopg2-binary==2.9.9 
pytz==2024.1